refactor: Move Booking controller as service, and code un routes to controllers
